requirements.txt
Django==5.0.1
djangorestframework==3.14.0
celery==5.3.6
django-celery-results==2.5.1
